pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "speech-to-text-app"
version = "0.1.0"
description = "Windows microphone dictation app backed by Google Cloud Speech-to-Text."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-cloud-speech>=2.24.0",
  "sounddevice>=0.4.7",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
